add lock function for student

+++ oa/src/main/java/com/hotmarzz/oa/buzz/impl/StudentBuzzImpl.java
@@ -9,12 +9,14 @@ import javax.servlet.http.HttpSession;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 import com.hotmarzz.basic.dao.BaseQuery;
 import com.hotmarzz.oa.buzz.StudentBuzz;
 import com.hotmarzz.oa.controller.EmpController;
 import com.hotmarzz.oa.dao.StudentDao;
+import com.hotmarzz.oa.exception.StudentLockException;
 import com.hotmarzz.oa.exception.StudentRepeatException;
 import com.hotmarzz.oa.pojo.Emp;
 import com.hotmarzz.oa.pojo.Student;
@@ -55,7 +57,12 @@ public class StudentBuzzImpl implements StudentBuzz {
 	}
 
 	@Override
-	public void update(Student stu) throws Exception {
+	public void update(Student stu) throws StudentLockException,Exception {
+		//检查不通过，抛出异常StudentLockException
+		checkLockPermission(stu.getStuId());
+		stu.setLocked(0);
+		stu.setLockTime(new Date());
+		stu.setLockUser(((Emp)session.getAttribute(SessionUtils.LOGIN_EMP_KEY)).getEmpName());
 		stuDao.update(stu);
 	}
 
@@ -84,6 +91,7 @@ public class StudentBuzzImpl implements StudentBuzz {
 	 */
 	@Override
 	public void delete(Long stuId) throws Exception {
+		checkLockPermission(stuId);
 		stuDao.delete(stuId);
 	}
 	
@@ -102,5 +110,27 @@ public class StudentBuzzImpl implements StudentBuzz {
 			return null;
 		}
 	}
-
+	/**
+	 * 检查指定ID的学生的锁定人，是否与当前登陆人一致。不一致会抛出
+	 * @param stuId
+	 * @return
+	 * @throws Exception
+	 */
+	private void checkLockPermission(Long stuId) throws StudentLockException,Exception{
+		Student stu = stuDao.getById(stuId);
+		String lockUserName =  ((Emp)session.getAttribute(SessionUtils.LOGIN_EMP_KEY)).getEmpName();
+		if(stu.getLocked() == 0 && !lockUserName.equals(stu.getLockUser())){
+			throw new StudentLockException(lockUserName);
+		}
+	}
+	
+	@Override
+	@Scheduled(cron = "0 0 0/1 * * ?")
+	public void updateUnlockForStus() throws Exception {
+		Date now = new Date();
+		Date before = new Date(now.getTime()-MILLI_SECOND_OF_UNLOCK_INTERVAL);
+		stuDao.unlockAllByLockTime(before);
+	}
+	
+	
 }
